XW-1713 | add base wiki-page handler, move Lyric page view function to ArticleHandler
