+++ front/main/app/utils/wiki-handlers/article.js
@@ -1,4 +1,7 @@
+import Ember from 'ember';
+import request from 'ember-ajax/request';
 import VisibilityStateManager from '../visibility-state-manager';
+import BaseHandler from './base-handler';
 
 /**
  * @param {Ember.Route} route
@@ -16,11 +19,50 @@ function afterModel(route, model) {
 }
 
 /**
+ * This function tracks page view only on articles on Lyrics Wiki (id: 43339)
+ *
+ * @param {Ember.Model} model
+ */
+function sendLyricsPageView(model) {
+	if (Ember.get(Mercury, 'wiki.id') === 43339) {
+		const amgId = parseInt($('#lyric').data('amg-id'), 10) || 0,
+			gracenoteId = parseInt($('#gracenoteid').text(), 10) || 0;
+
+		request(M.buildUrl({path: '/wikia.php'}), {
+			data: {
+				controller: 'LyricFind',
+				method: 'track',
+				title: model.title,
+				amgid: amgId,
+				gracenoteid: gracenoteId,
+				rand: (`${Math.random()}`).substr(2, 8)
+			},
+			dataType: 'text'
+		}).catch((error) => {
+			/**
+			 * MediaWiki returns 404 with header X-LyricFind-API-Code:106
+			 * for success request but no lyrics
+			 */
+		});
+	}
+}
+
+/**
+ * Hook triggered on transition.then() in Route::afterModel()
+ *
+ * @param model
+ */
+function afterTransition(model) {
+	sendLyricsPageView(model);
+}
+
+/**
  * Export Article handler
  */
-export default {
+export default Ember.$.extend({}, BaseHandler, {
 	// template's and controller's name
 	controllerName: 'article',
 	viewName: 'article',
-	afterModel
-};
+	afterModel,
+	afterTransition
+});
