+++ front/main/app/utils/wiki-handlers/base-handler.js
@@ -1,0 +1,9 @@
+/**
+ * This object acts like the interface, but implementation of fields ond methods is optional
+ */
+export default {
+	controllerName: '',
+	viewName: '',
+	afterModel() {},
+	afterTransition() {}
+};
