+++ front/main/app/utils/wiki-handlers/category.js
@@ -1,4 +1,5 @@
 import ArticleHandler from './article';
+import BaseHandler from './base-handler';
 
 /**
  * afterModel hook
@@ -14,10 +15,10 @@ function afterModel(route, model) {
 /**
  * Export Category handler
  */
-export default {
+export default Ember.$.extend({}, BaseHandler, {
 	// template's and controller's name
 	viewName: 'category',
 	controllerName: 'category',
 	// hooks
-	afterModel,
-};
+	afterModel
+});
