+++ front/main/app/utils/wiki-handlers/curated-main-page.js
@@ -1,5 +1,6 @@
 import Ember from 'ember';
 import CuratedContentModel from '../../models/curated-content';
+import BaseHandler from './base-handler';
 
 /**
  * Set curatedContent data if main page has curated content set
@@ -40,12 +41,10 @@ function afterModel(route, model) {
 	return model;
 }
 
-export default {
+export default Ember.$.extend({}, BaseHandler, {
 	// template's and controller's name
 	controllerName: 'main-page',
 	viewName: 'main-page',
 	// hooks
-	afterModel,
-	// all other, handler-specific functions
-	getCuratedContentModel,
-};
+	afterModel
+});
